+++ Sources/SwiftUIKit/_Deprecated/ActionButton.swift
@@ -1,0 +1,6 @@
+@available(iOS 14, *)
+@available(*, deprecated, renamed: "FormActionButton")
+public typealias ActionButton = FormActionButton
+
+@available(*, deprecated, renamed: "FormActionButtonStyle")
+public typealias ActionButtonStyle = FormActionButtonStyle